Atomic file store resets properly

+++ Core/Lokad.Cqrs.Portable/AtomicStorage/FileAtomicStorageFactory.cs
@@ -78,7 +78,8 @@ namespace Lokad.Cqrs.AtomicStorage
 
         public void Reset()
         {
-            Directory.Delete(_folderPath, true);
+            if (Directory.Exists(_folderPath))
+                Directory.Delete(_folderPath, true);
             Directory.CreateDirectory(_folderPath);
         }
     }
